+++ csatar-plugins/csatar/lang/hu/lang.php
@@ -52,6 +52,7 @@
                 'inactivationWarning' => 'Figyelem! Ha a státusz aktívról bármilyen típusú inaktív státuszra változik, a szervezeti egység alá tartozó összes szervezeti egység és cserkész státusza is inaktívvá válik, valamint lejár az ezekhez tartozó összes mebízatás!',
                 'bulkDeletePartialSuccess' => ':deletedCount elem törlése sikerült a :totalCount-ból/ből. ',
                 'bulkDeleteError' => 'A következő hibák merültek fel a törlés során: ',
+                'activeMembersCount' => 'Aktív tagok száma',
             ],
             'ageGroups' => [
                 'ageGroups' => 'Korosztályok',