+++ csatar-plugins/csatar/models/Patrol.php
@@ -27,6 +27,8 @@ class Patrol extends OrganizationBase
 
     protected $appends = ['extended_name'];
 
+    public $customAttributes = ['active_members_count'];
+
     /**
      * @var array Validation rules
      */
@@ -333,5 +335,9 @@ class Patrol extends OrganizationBase
     public function scopeActive($query)
     {
         return $query->where('status', Status::ACTIVE);
+    }
+
+    public function getActiveMembersCountAttribute() {
+        return StructureTree::getPatrolScoutsCount($this->id);
     }
 }