+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- CS-493 - Added active members count to Association, District, Team, Troop, Patrol models
+- CS-493 - Modified form builder to display custom attributes
+- CS-493 - Added missing translations
 - CS-411 - Fixed the header on the team page, so neither the image, nor the header will overflow the screen
 - CS-387 - Created Game development front end menu, added Create New Game and Edit Game buttons, added functionality to approve game
 - CS-387 - Created Partials component to make partial sharing possible between plugins